+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All Notable changes to `bytenet\laravel-admn-base` project will be documented in
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2016-12-29
+### Added
+- src/day04/README.md
+- src/day04/PlayerMustPickFirstException.php
+- src/day04/MontyHallStrategy.php
+- src/day04/MontyHall.php
+- tests/day04/MontyHallTest.php
+
 ## [0.3.0] - 2016-12-28
 ### Added
 - src/day03/README.md
@@ -36,6 +44,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 -  Day-01 - SRC and TEST
 
+[0.4.0]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.1.0...v0.1.1